Client refactor
